Fix potential bug with ColorNonOverworldSprites function.
Register "b" was supposed to be a counter, but then it was used for
something else. It didn't really cause problems because it halted on the
first unused sprite, but it could have been an issue if all 40 sprites
were in use.

For that matter, the fact that it was stopping at the first zero'd out
sprite was bad behaviour as well, because sometimes slots are skipped
(ie. when "cut" is used in the overworld, though that didn't matter in
this case). Anyway, now it checks all 40 sprites every time.