Optimized "WindowTransferBgRowsAndColors".
This was the only function that was still copied directly from
froggest's code. It had hblank wait loops - which is bad since it's
called from vblank, so it took far longer than it needed to, and
everything after that in the vblank routine failed.

The only effect was that "OAM DMA" occurred in the middle of a scanline.
This caused artifacts on a real gameboy when opening menus, but
emulators didn't seem to care at all.

Also made "OAM DMA" call go before palette-loading routines, just in
case. This broke VBA again.