Fix bug with the CopyVideoData function.
If vblank occurs while DelayFrameHook is executing its "extra
functions", that's a problem, because certain functions (such as
"CopyVideoData") expect that the ROM bank won't be changed when calling
DelayFrame. The fix is to save the ROM bank in DelayFrame, and have the
InterruptWrapper restore it if the interrupt occurs at that time.

Before, the extra function calls in DelayFrameHook were wrapped in
a di/ei block, but that was removed since it caused problems with
delaying vblank.

This manifested itself in graphical corruption while healing pokemon,
but it was very inconsistent, occurring on some savefiles but not
others.