Fixed the ghost marowak reveal animation.
I told the "SetPal_Battle" function to stop loading sprite palettes. The
only reason it did that was for the pokeball palettes. Instead, I hooked
the "LoadPartyPokeballGfx" function and and told it to load the palettes
then.

This was all necessary to prevent SetPal_Battle from overwriting palette
changes needed for the marowak animation.